+++ src/components/accession2/edit/CalculatorModal.tsx
@@ -5,7 +5,7 @@ import DialogBox from 'src/components/common/DialogBox/DialogBox';
 import { Box, Grid, Theme } from '@mui/material';
 import { Textfield } from '@terraware/web-components';
 import { Accession } from 'src/types/Accession';
-import AccessionsService from 'src/services/AccessionsService';
+import AccessionService from 'src/services/AccessionService';
 import { Unit, usePreferredWeightUnits } from 'src/units';
 import useSnackbar from 'src/utils/useSnackbar';
 import { Dropdown } from '@terraware/web-components';
@@ -48,7 +48,7 @@ export default function CalculatorModal(props: CalculatorModalProps): JSX.Elemen
 
   const getTotalCount = async () => {
     if (validateFields()) {
-      const response = await AccessionsService.updateAccession(record, true);
+      const response = await AccessionService.updateAccession(record, true);
       if (response.requestSucceeded && response.accession) {
         if (
           response.accession.subsetWeight?.grams &&
