+++ src/components/accession2/edit/EditState.tsx
@@ -3,7 +3,7 @@ import strings from 'src/strings';
 import { Box, Grid, Theme, Typography, useTheme } from '@mui/material';
 import { Dropdown, Icon } from '@terraware/web-components';
 import { Accession, AccessionState, stateName } from 'src/types/Accession';
-import { AccessionsService } from 'src/services';
+import { AccessionService } from 'src/services';
 import { makeStyles } from '@mui/styles';
 
 const useStyles = makeStyles((theme: Theme) => ({
@@ -25,7 +25,7 @@ export default function EditState(props: EditStateProps): JSX.Element {
   const { accession, record, onChange, error } = props;
 
   const stateChanged = accession.state !== record.state;
-  const options = AccessionsService.getTransitionToStates(accession.state).map((state: AccessionState) => ({
+  const options = AccessionService.getTransitionToStates(accession.state).map((state: AccessionState) => ({
     label: stateName(state),
     value: state,
   }));
