+++ src/components/accession2/properties/CollectedReceivedDate2.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useState, useEffect } from 'react';
 import { useTheme, Grid } from '@mui/material';
 import { DatePicker } from '@terraware/web-components';
 import { Accession } from 'src/types/Accession';
-import { AccessionPostRequestBody } from 'src/services/AccessionsService';
+import { AccessionPostRequestBody } from 'src/services/SeedBankService';
 import getDateDisplayValue, { isInTheFuture } from '@terraware/web-components/utils/date';
 import strings from 'src/strings';
